requirements.txt
Django>=3.0,<4.0
celery>=5.0.0
flower
djangorestframework==3.12.2
